GenAI-Project/requirements.txt
PyMuPDF
pandas
numpy
git+https://github.com/facebookresearch/nougat
pypdfium2
transformers
torch
accelerate
bitsandbytes
tqdm
langchain
langchain-huggingface
accelerate
bitsandbytes
pydantic
boto3
botocore
pyyaml
